+++ PndMCMatch/PndMCObject.cxx
@@ -31,7 +31,7 @@ void PndMCObject::SetEntry(std::vector<Int_t> type, std::vector<Int_t> link, int
 	PndMCEntry myEntry;
 	myEntry.SetPos(index);
 	myEntry.SetPersistanceCheck(kFALSE);
-	for (int i = 0; i < type.size(); i++){
+	for (size_t i = 0; i < type.size(); i++){
 		myEntry.AddLink(FairLink(type[i],link[i]));
 	}
 	SetEntry(myEntry);
@@ -45,7 +45,7 @@ void PndMCObject::SetEntry(FairMultiLinkedData* data, int index)
 
 void PndMCObject::SetLink(FairLink link, int index)
 {
-	if (index < fStage.size()){
+	if (index < (int)fStage.size()){
 		fStage[index].ResetLinks();
 	}
 	AddLink(link, index);
@@ -61,7 +61,7 @@ void PndMCObject::AddLink(FairLink link, int index)
 
 void PndMCObject::AdoptSize(int index){
 	int start = fStage.size();
-	while (fStage.size() < index+1){
+	while ((int)fStage.size() < index+1){
 		PndMCEntry myVec;
 		myVec.SetPos(fStage.size());
 		myVec.SetSource(GetStageId());
@@ -87,7 +87,7 @@ PndMCEntry PndMCObject::GetEntry(int index) const
 FairMultiLinkedData PndMCObject::PosInList(FairLink link){
 	FairMultiLinkedData result;
 	result.SetPersistanceCheck(kFALSE);
-	for (int i = 0; i < fStage.size(); i++){
+	for (size_t i = 0; i < fStage.size(); i++){
 		if (fStage[i].IsLinkInList(link.GetType(), link.GetIndex()))
 			result.AddLink(FairLink(GetStageId(), i));
 	}
@@ -96,7 +96,7 @@ FairMultiLinkedData PndMCObject::PosInList(FairLink link){
 
 void PndMCObject::RemoveType(Int_t type)
 {
-	for (int i = 0; i < fStage.size(); i++){
+	for (size_t i = 0; i < fStage.size(); i++){
 		fStage[i].RemoveType(type);
 	}
 }
